Created dashboard blueprint and added home route to it
