+++ api/templates/base.html
@@ -17,7 +17,7 @@
     <header class="base-header">
       <a href="{{ url_for('info.index') }}" class="logo">QUIZIVERSE</a>
       <nav>
-        <a href="{{ url_for('home') }}">Home</a>
+        <a href="{{ url_for('dash.home') }}">Home</a>
         <a href="{{ url_for('info.about') }}">About</a>
         <a href="{{ url_for('auth.login') }}">Login</a>
       </nav>
