+++ api/templates/create.html
@@ -46,7 +46,7 @@
         <button type="button" id="add-question-btn">Add</button>
         <button type="button" id="remove-question-btn">Remove</button><br>
         <button type="submit" class="submit-btn">Finish</button>
-        <a href="{{ url_for('home') }}" class="cancel-link"><button type="button">Cancel</button></a>
+        <a href="{{ url_for('dash.home') }}" class="cancel-link"><button type="button">Cancel</button></a>
       </div>
     </form>
     
@@ -83,7 +83,7 @@
         <button type="button" id="add-question-btn">Add</button>
         <button type="button" id="remove-question-btn">Remove</button><br>
         <button type="submit" class="submit-btn">Finish</button>
-        <a href="{{ url_for('home') }}" class="cancel-link"><button type="button">Cancel</button></a>
+        <a href="{{ url_for('dash.home') }}" class="cancel-link"><button type="button">Cancel</button></a>
       </div>
 
     </form>
