+++ api/templates/finishquiz.html
@@ -30,7 +30,7 @@
     </div>
     
     <div class="results-actions">
-      <a href="{{ url_for('home') }}" class="button return-home">Return Home</a>
+      <a href="{{ url_for('dash.home') }}" class="button return-home">Return Home</a>
       <a href="{{ url_for('quizinfo', quiz_id=quiz_id) }}" class="button retry-quiz">Retry</a>
     </div>
   </div>
